Implement common base for TMDB cogs
